+++ src/lib/types/google-sheet/google-sheet-opts.type.ts
@@ -1,0 +1,5 @@
+export interface GoogleSheetsOpts {
+    sheetId: string;
+    clientEmail: string;
+    privateKey: string;
+}