+++ src/lib/types/google-sheet/temporal-data.type.ts
@@ -1,0 +1,13 @@
+export interface TemporalDataType {
+    date?: string;
+    time?: string;
+    name?: string;
+    phone?: string;
+    quantity?: string;
+}
+
+export enum TemporalStatusEnum {
+    JUST_STARTED = 'JUST_STARTED',
+    IN_PROGRESS = 'IN_PROGRESS',
+    COMPLETED = 'COMPLETED',
+}