+++ src/modules/google-sheets/helpers/temporal-data.helper.ts
@@ -1,0 +1,12 @@
+import { TemporalDataType } from "src/lib";
+
+export function customDataSheetHelper(values: TemporalDataType): TemporalDataType {
+    return {
+        date: values.date ?? " ",
+        time: values.time ?? " ",
+        name: values.name ?? " ",
+        phone: values.phone ?? " ",
+        quantity: values.quantity ?? " ",
+      };
+}
+